# ZelBack/src/services/appManagement/operationsController.js
'use strict';

const messageHelper = require('../messageHelper');
const serviceHelper = require('../serviceHelper');
const verificationHelper = require('../verificationHelper');
const jobRegistry = require('../utils/jobRegistry');
const fluxNetworkHelper = require('../fluxNetworkHelper');
const { nodeApiUrl } = require('../utils/socketAddressUtils');
const log = require('../../lib/log');
const { Privilege, authOf } = require('../utils/privileges');

// The one status resource for every long-running operation this node accepts.
// Endpoints that start work answer 202 with a job handle and point here; this
// owns the polling contract so no endpoint has to invent one.

/**
 * A status path as the caller must actually use it.
 *
 * The registry deals in paths (`/apps/operations/<id>`) because a job does not
 * know which node it is on. A caller does need to know: an operation is
 * node-local, so a poll that lands anywhere else answers 404, and the address
 * the client happened to use may have been the load balancer rather than this
 * node. Absolute against this node's own DNS name closes both.
 *
 * Relative is the fallback when the node cannot resolve its own address — a
 * caller that already reached this node can still follow it, which is strictly
 * better than emitting a URL built from an address we could not confirm.
 *
 * @param {string} statusPath
 * @returns {Promise<string>}
 */
async function absoluteStatusUrl(statusPath) {
  try {
    const base = nodeApiUrl(await fluxNetworkHelper.getLocalSocketAddress());
    return base ? `${base}${statusPath}` : statusPath;
  } catch (error) {
    log.warn(`operationsController: could not resolve this node's address for a status URL: ${error.message}`);
    return statusPath;
  }
}

/**
 * The FluxID a caller is authenticated as, or null. Jobs registered with an
 * owner are only readable by that identity; jobs registered without one treat
 * the jobId itself as the capability.
 */
async function callerFluxId(req) {
  const authorized = await verificationHelper.verifyPrivilege(Privilege.USER, authOf(req));
  if (!authorized) return null;
  const auth = serviceHelper.ensureObject(authOf(req));
  return auth ? auth.zelid : null;
}

/**
 * Shape a 202 for an endpoint that has just started work. Location and
 * Operation-Id are the RFC 9110 / Azure long-running-operation spelling; the
 * body repeats them so a client that cannot read headers is not stuck.
 *
 * @param {import('express').Response} res
 * @param {{jobId: string, statusUrl: string}} handle
 * @param {object} [extra] additional body fields the endpoint wants echoed
 */
async function accepted(res, handle, extra = {}) {
  const statusUrl = await absoluteStatusUrl(handle.statusUrl);
  res.setHeader('Location', statusUrl);
  res.setHeader('Operation-Id', handle.jobId);
  res.setHeader('Retry-After', String(jobRegistry.retryAfterSeconds()));
  return res.status(202).json(messageHelper.createDataMessage({
    jobId: handle.jobId,
    statusUrl,
    status: jobRegistry.JobStatus.RUNNING,
    ...extra,
  }));
}

/**
 * How far through an operation's line-numbered output the caller has already
 * read. Anything that is not a non-negative whole number is treated as no
 * cursor at all - a client sending nonsense gets the whole retained view rather
 * than a silently truncated one.
 *
 * @param {import('express').Request} req
 * @returns {number} 0 when absent or unusable
 */
/**
 * Answer an operation that finished while the caller was still waiting.
 *
 * 200 rather than 202, because there is nothing to come back for: the job is in
 * its terminal state and the body carries it. A client that predates jobs sees
 * what it has always seen - a completed request - and one that knows about them
 * reads the same status it would have polled for.
 *
 * @param {object} res
 * @param {{jobId: string, statusUrl: string}} handle
 * @returns {object}
 */
function completed(res, handle, owner = null) {
  res.setHeader('Operation-Id', handle.jobId);

  // The owner has to be given, because a job carries one and the registry
  // refuses a read that does not match it. Read without it, every owned job
  // came back null and the answer fell through to a hardcoded Succeeded - so an
  // operation that FAILED was reported to its caller as having worked.
  const job = jobRegistry.get(handle.jobId, owner);

  // Gone or unreadable is not the same as succeeded. The caller is told the
  // work is over, and is not told that it worked.
  if (!job) {
    return res.status(200).json(messageHelper.createErrorMessage(
      'The operation finished but its result could not be read',
    ));
  }

  if (job.status !== jobRegistry.JobStatus.SUCCEEDED) {
    // message, name and code, which is the shape a failed file operation has
    // always answered with and what the dashboards read. The problem document
    // is built from an Error, so its title IS that error's name.
    const problem = job.error || {};
    return res.status(200).json(messageHelper.createErrorMessage(
      problem.detail || problem.title || `Operation ${job.status}`,
      problem.title,
      problem.code,
    ));
  }

  return res.status(200).json(messageHelper.createDataMessage({
    jobId: handle.jobId,
    statusUrl: handle.statusUrl,
    status: job.status,
  }));
}

function readCursor(req) {
  const raw = req.query && req.query.sinceSeq;
  if (raw === undefined || raw === null || raw === '') return 0;
  const seq = Number(raw);
  if (!Number.isSafeInteger(seq) || seq < 0) return 0;
  return seq;
}

/**
 * @param {import('express').Request} req
 * @param {import('express').Response} res
 */
/**
 * An operation's view for a caller, or null when there is nothing they may see.
 *
 * No req and no res: this is what BOTH surfaces read an operation through, and
 * each formats the answer in its own envelope. Unknown, expired and not-yours
 * are deliberately one answer - a jobId must not tell a caller whether someone
 * else has an operation running.
 *
 * @param {object} params
 * @param {string} params.jobId
 * @param {string|null} [params.callerId] the FluxID the caller is authenticated as
 * @param {number|null} [params.sinceSeq] progress cursor
 * @returns {Promise<object|null>}
 */
async function readOperation({ jobId, callerId = null, sinceSeq = null }) {
  const view = jobRegistry.get(jobId, callerId, { sinceSeq });
  if (!view) return null;
  // The problem-detail `instance` is the same status resource under another
  // name, so it is absolute for the same reason.
  if (view.error && view.error.instance) {
    return { ...view, error: { ...view.error, instance: await absoluteStatusUrl(view.error.instance) } };
  }
  return view;
}

async function getOperation(req, res) {
  try {
    const jobId = req.params.jobId || (req.query && req.query.jobId);
    if (!jobId) {
      return res.status(400).json(messageHelper.createErrorMessage('Missing jobId'));
    }

    const view = await readOperation({
      jobId, callerId: await callerFluxId(req), sinceSeq: readCursor(req),
    });
    if (!view) {
      return res.status(404).json(messageHelper.createErrorMessage('Operation not found'));
    }

    // A running operation is a 200 with a non-terminal body. Completion is read
    // from the status field, never inferred from the HTTP code - a failed
    // operation is still a successful poll.
    if (jobRegistry.isTerminal(view.status)) {
      res.setHeader('Expires', new Date(view.lastUpdatedAt + 60 * 60 * 1000).toUTCString());
    } else {
      res.setHeader('Retry-After', String(jobRegistry.retryAfterSeconds()));
    }

    return res.json(messageHelper.createDataMessage(view));
  } catch (error) {
    log.error(`operationsController getOperation: ${error.message}`);
    return res.status(500).json(messageHelper.createErrorMessage(error.message));
  }
}

/**
 * @param {import('express').Request} req
 * @param {import('express').Response} res
 */
async function cancelOperation(req, res) {
  try {
    const { jobId } = req.params;
    if (!jobId) {
      return res.status(400).json(messageHelper.createErrorMessage('Missing jobId'));
    }

    const owner = await callerFluxId(req);
    const view = jobRegistry.get(jobId, owner);
    if (!view) {
      return res.status(404).json(messageHelper.createErrorMessage('Operation not found'));
    }

    // Best effort, and said so plainly: the flag is raised here and the worker
    // stops at its next checkpoint, so the status stays Running until it does.
    const requested = jobRegistry.requestCancel(jobId);
    return res.json(messageHelper.createDataMessage({
      jobId,
      cancelRequested: requested,
      status: jobRegistry.get(jobId, owner).status,
    }));
  } catch (error) {
    log.error(`operationsController cancelOperation: ${error.message}`);
    return res.status(500).json(messageHelper.createErrorMessage(error.message));
  }
}

module.exports = {
  readOperation,
  callerFluxId,
  completed,
  absoluteStatusUrl,
  accepted,
  getOperation,
  cancelOperation,
};